Run closure compiler inside a compilation worker.
