+++ package.json
@@ -166,6 +166,7 @@
     "nock": "13.1.4",
     "node-fetch": "2.6.5",
     "open": "8.4.0",
+    "piscina": "3.1.0",
     "plugin-error": "1.0.1",
     "postcss": "8.3.11",
     "postcss-import": "14.0.2",
